+++ internal/workertest/workerpool.go
@@ -1,4 +1,4 @@
-package workers
+package workertest
 
 import (
 	"context"
@@ -10,6 +10,7 @@ import (
 
 	"github.com/temporalio/omes/clioptions"
 	"github.com/temporalio/omes/loadgen"
+	"github.com/temporalio/omes/workers"
 	"go.uber.org/zap"
 )
 
@@ -52,7 +53,7 @@ func (w *workerPool) ensureWorkerBuilt(
 	w.mutex.Unlock()
 
 	once.Do(func() {
-		baseDir := BaseDir(w.env.repoDir, sdk)
+		baseDir := workers.BaseDir(w.env.repoDir, sdk)
 		buildDir := filepath.Join(baseDir, w.env.buildDirName())
 
 		w.mutex.Lock()
@@ -63,7 +64,7 @@ func (w *workerPool) ensureWorkerBuilt(
 		})
 		w.mutex.Unlock()
 
-		builder := Builder{
+		builder := workers.Builder{
 			DirName:    w.env.buildDirName(),
 			SdkOptions: clioptions.SdkOptions{Language: sdk},
 			Logger:     logger.Named(fmt.Sprintf("%s-builder", sdk)),
@@ -97,9 +98,9 @@ func (w *workerPool) startWorker(
 
 	go func() {
 		defer close(workerDone)
-		baseDir := BaseDir(w.env.repoDir, sdk)
-		runner := &Runner{
-			Builder: Builder{
+		baseDir := workers.BaseDir(w.env.repoDir, sdk)
+		runner := &workers.Runner{
+			Builder: workers.Builder{
 				DirName:    w.env.buildDirName(),
 				SdkOptions: clioptions.SdkOptions{Language: sdk},
 				Logger:     logger.Named(fmt.Sprintf("%s-worker-builder", sdk)),
